apps/api/requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
sqlmodel==0.0.35
psycopg[binary]==3.2.12
pydantic-settings==2.13.1
python-dotenv==1.2.1
email-validator==2.3.0
passlib==1.7.4
python-jose==3.5.0
bcrypt==5.0.0
cryptography==46.0.5
httpx==0.28.1
feedparser==6.0.12
python-dateutil==2.9.0.post0
tzdata==2025.3

